Extracted status to common
