+++ src/common/status/component.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import styled, { css } from "styled-components";
-import { FONT_SIZE_DATA, FONT_SIZE_ENUM } from "../../../theme/sizes";
-import { STATUS_COLOR_DATA } from "../../../theme/color";
+import { FONT_SIZE_DATA, FONT_SIZE_ENUM } from "../../theme/sizes";
+import {STATUS_COLOR_DATA} from "../../theme/color";
 
 const Status: React.FC<{ message?: string }> = ({ message }) => {
   return (
@@ -20,11 +20,11 @@ const StyledContainer = styled.div<{ message: string }>`
   border-radius: 10px;
   padding: 10px ;
   ${({message}) => {
-    return css`
+  return css`
       border-color: ${message === "Successful!" ? STATUS_COLOR_DATA["success"] : STATUS_COLOR_DATA["error"]};
       color: ${message === "Successful!" ? STATUS_COLOR_DATA["success"] : STATUS_COLOR_DATA["error"]};
     `;
-  }}
+}}
 `;
 
 export default Status;
