+++ src/common/status/index.tsx
@@ -1,0 +1,3 @@
+import Component from "./component";
+
+export {Component as Status}
